TlsDecorator: fix deadlock issues
Worker thread should not call tls_read, tls_write,
or any method of the lower layer connection while holding
the decorator's mutex, because the lower layer may have
(and does have, in the case of NetworkConnection) its own
worker thread which calls delegates while holding its own
mutex.

We don't need receiveBufferDecrypted to be in the object
state, and it doesn't need to be protected by any mutex,
because it's only needed by the worker thread.  We also
need a temporary buffer to hold data to be sent to tls_write.
So move receiveBufferDecrpyted to the worker thread and
just call it "buffer".  Reuse the buffer for tls_write.